Rename exthm, solthm
ex and sol were renamed to exthm and solthm and then wrapped in a new
environment in 54f7b4314c281c5c22afb083311cbab2458e5a35. But this
changed the URLs to be exthm.x instead of ex.x, etc., which is ugly.

Here we rename them, so URLs become exercise.x.